LMFAO

+++ models.py
@@ -84,7 +84,7 @@ class Generator(nn.Module):
         outputs = torch.zeros(batch_size, max_len, vocab_size).cuda() #, requires_grad=False)
         max_output = torch.zeros(batch_size, max_len).cuda() #, requires_grad=False)
 
-        decoder_input = torch.zeros(batch_size, dtype=torch.long).cuda() #, requires_grad=False)
+        decoder_input = torch.zeros(batch_size, dtype=torch.int32).cuda() #, requires_grad=False)
         max_output[:,0] = decoder_input
 
         for t in range(max_len):
